+++ packages/components/base/source/1-atoms/rich-text/index.ts
@@ -1,4 +1,5 @@
-import { forwardRef, createElement, createContext, useContext } from 'react';
+import { forwardRef, createContext } from 'react';
+import { withContainer } from '@kickstartds/core/lib/container';
 import {
   RichTextComponent,
   type RichTextProps,
@@ -9,6 +10,4 @@ import './rich-text.scss';
 export { RichTextProps, defaultRenderFn };
 export const RichTextContextDefault = forwardRef(RichTextComponent);
 export const RichTextContext = createContext(RichTextContextDefault);
-export const RichText: typeof RichTextContextDefault = forwardRef(
-  (props, ref) => createElement(useContext(RichTextContext), { ...props, ref })
-);
+export const RichText = withContainer('rich-text', RichTextContext);
