+++ packages/components/base/source/1-atoms/table/index.ts
@@ -1,4 +1,5 @@
-import { forwardRef, createElement, createContext, useContext } from 'react';
+import { forwardRef, createContext } from 'react';
+import { withContainer } from '@kickstartds/core/lib/container';
 import { TableComponent, type TableProps } from './TableComponent';
 import './table.scss';
 import './ResponsiveTable.js';
@@ -6,6 +7,4 @@ import './ResponsiveTable.js';
 export { TableProps };
 export const TableContextDefault = forwardRef(TableComponent);
 export const TableContext = createContext(TableContextDefault);
-export const Table: typeof TableContextDefault = forwardRef((props, ref) =>
-  createElement(useContext(TableContext), { ...props, ref })
-);
+export const Table = withContainer('table', TableContext);
