+++ packages/components/base/source/2-molecules/slider/index.ts
@@ -1,4 +1,5 @@
-import { forwardRef, createElement, createContext, useContext } from 'react';
+import { forwardRef, createContext } from 'react';
+import { withContainer } from '@kickstartds/core/lib/container';
 import { SliderComponent, type SliderProps } from './SliderComponent';
 import './slider.scss';
 import './lazySlider.js';
@@ -7,6 +8,4 @@ export { SliderProps };
 export { SlideContext } from './SlideContext';
 export const SliderContextDefault = forwardRef(SliderComponent);
 export const SliderContext = createContext(SliderContextDefault);
-export const Slider: typeof SliderContextDefault = forwardRef((props, ref) =>
-  createElement(useContext(SliderContext), { ...props, ref })
-);
+export const Slider = withContainer('slider', SliderContext);
