+++ packages/components/base/source/2-molecules/teaser-box/index.ts
@@ -1,10 +1,9 @@
-import { forwardRef, createElement, createContext, useContext } from 'react';
+import { forwardRef, createContext } from 'react';
+import { withContainer } from '@kickstartds/core/lib/container';
 import { TeaserBoxComponent, type TeaserBoxProps } from './TeaserBoxComponent';
 import './teaser-box.scss';
 
 export { TeaserBoxProps };
 export const TeaserBoxContextDefault = forwardRef(TeaserBoxComponent);
 export const TeaserBoxContext = createContext(TeaserBoxContextDefault);
-export const TeaserBox: typeof TeaserBoxContextDefault = forwardRef(
-  (props, ref) => createElement(useContext(TeaserBoxContext), { ...props, ref })
-);
+export const TeaserBox = withContainer('teaser-box', TeaserBoxContext);
