+++ packages/components/base/source/2-molecules/text-media/index.ts
@@ -1,10 +1,9 @@
-import { forwardRef, createElement, createContext, useContext } from 'react';
+import { forwardRef, createContext } from 'react';
+import { withContainer } from '@kickstartds/core/lib/container';
 import { TextMediaComponent, type TextMediaProps } from './TextMediaComponent';
 import './text-media.scss';
 
 export { TextMediaProps };
 export const TextMediaContextDefault = forwardRef(TextMediaComponent);
 export const TextMediaContext = createContext(TextMediaContextDefault);
-export const TextMedia: typeof TextMediaContextDefault = forwardRef(
-  (props, ref) => createElement(useContext(TextMediaContext), { ...props, ref })
-);
+export const TextMedia = withContainer('text-media', TextMediaContext);
