+++ packages/components/content/source/2-molecules/logo-tiles/index.ts
@@ -1,10 +1,12 @@
-import { forwardRef, createElement, createContext, useContext } from 'react';
+import { forwardRef, createContext } from 'react';
+import { withContainer } from '@kickstartds/core/lib/container';
 import { LogoTilesComponent, type LogoTilesProps } from './LogoTilesComponent';
 import './logo-tiles.scss';
 
 export { LogoTilesProps };
 export const LogoTilesContextDefault = forwardRef(LogoTilesComponent);
 export const LogoTilesContext = createContext(LogoTilesContextDefault);
-export const LogoTiles: typeof LogoTilesContextDefault = forwardRef(
-  (props, ref) => createElement(useContext(LogoTilesContext), { ...props, ref })
+export const LogoTiles: typeof LogoTilesContextDefault = withContainer(
+  'logo-tiles',
+  LogoTilesContext
 );
