+++ packages/components/content/source/2-molecules/quote/index.ts
@@ -1,10 +1,9 @@
-import { forwardRef, createElement, createContext, useContext } from 'react';
+import { forwardRef, createContext } from 'react';
+import { withContainer } from '@kickstartds/core/lib/container';
 import { QuoteComponent, type QuoteProps } from './QuoteComponent';
 import './quote.scss';
 
 export { QuoteProps };
 export const QuoteContextDefault = forwardRef(QuoteComponent);
 export const QuoteContext = createContext(QuoteContextDefault);
-export const Quote: typeof QuoteContextDefault = forwardRef((props, ref) =>
-  createElement(useContext(QuoteContext), { ...props, ref })
-);
+export const Quote = withContainer('quote', QuoteContext);
