+++ packages/components/content/source/2-molecules/visual/VisualComponent.tsx
@@ -1,11 +1,5 @@
-import {
-  ForwardRefRenderFunction,
-  HTMLAttributes,
-  forwardRef,
-  createContext,
-} from 'react';
+import { ForwardRefRenderFunction, HTMLAttributes } from 'react';
 import classnames from 'classnames';
-import { withContainer } from '@kickstartds/core/lib/container';
 import { Icon } from '@kickstartds/base/lib/icon';
 import { type VisualProps as VisualSchemaProps } from './VisualProps';
 import { VisualMediaPartial } from './VisualMediaPartial';
@@ -59,7 +53,3 @@ export const VisualComponent: ForwardRefRenderFunction<
     )}
   </div>
 );
-
-export const VisualContextDefault = forwardRef(VisualComponent);
-export const VisualContext = createContext(VisualContextDefault);
-export const Visual = withContainer('visual', VisualContext);
