+++ packages/components/content/source/2-molecules/visual/index.ts
@@ -1,4 +1,5 @@
-import { forwardRef, createElement, createContext, useContext } from 'react';
+import { forwardRef, createContext } from 'react';
+import { withContainer } from '@kickstartds/core/lib/container';
 import { VisualComponent, type VisualProps } from './VisualComponent';
 import './visual.scss';
 import './Visual.js';
@@ -6,6 +7,4 @@ import './Visual.js';
 export { VisualProps };
 export const VisualContextDefault = forwardRef(VisualComponent);
 export const VisualContext = createContext(VisualContextDefault);
-export const Visual: typeof VisualContextDefault = forwardRef((props, ref) =>
-  createElement(useContext(VisualContext), { ...props, ref })
-);
+export const Visual = withContainer('visual', VisualContext);
